+++ models/Order.ts
@@ -1,5 +1,5 @@
 import {Schema, model, Types} from 'mongoose'
-import {Product} from "./Product.js"
+import {Product} from "./Product.ts"
 
 const schema = new Schema({
     user: {type: Types.ObjectId, ref: 'User'},
